added numbers print

+++ printfuncs/ft_putnbr.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void		ft_putnbr(int n)
+void		ft_putnbr(long n)
 {
 	ft_putnbr_fd(n, 1);
 }
